Fixing tests and lint
